delete the old FullScreenMemory widget and file and rename new FullScreenMemoryNew widget and file to 'FullScreenMemory'
