+++ lib/ui/home/memories/memory_cover_widget.dart
@@ -2,7 +2,7 @@ import "package:flutter/material.dart";
 import "package:photos/generated/l10n.dart";
 import "package:photos/models/memory.dart";
 import "package:photos/theme/ente_theme.dart";
-import "package:photos/ui/home/memories/full_screen_memory_new.dart";
+import 'package:photos/ui/home/memories/full_screen_memory.dart';
 import "package:photos/ui/viewer/file/thumbnail_widget.dart";
 import "package:photos/utils/navigation_util.dart";
 
@@ -35,7 +35,7 @@ class _MemoryCovertWidgetState extends State<MemoryCovertWidget> {
           FullScreenMemoryDataUpdater(
             initialIndex: index,
             memories: widget.memories,
-            child: FullScreenMemoryNew(title, index),
+            child: FullScreenMemory(title, index),
           ),
           forceCustomPageRoute: true,
         );
